requirements.txt
openai>=1.0.0
pymysql
python-dotenv
